Cryptography
~~~~~~~~~~~~~~~~~~~~~~
Cryptography -- The Caesar Cipher
~~~~~~~~~~~~~~~~~~~~~~

Your task for this week is to develop a simple Caesar Cipher. A Caesar Cipher is a cipher in which each letter maps to another letter of the alphabet based on a shift.

Suppose you had two strips of paper with the alphabet written out on each. If you align them so that A pairs with A, B with B, C with C, etc and then shift the top paper five letters to the right, you have a Caesar Cipher of +5.

The top paper's A would match with the bottom's F, the top's B with the bottom's G, the top's C with the bottom's H and so on until the top's Z pairs with the bottom's E (making a full rotation). (See attached image for an example.)

Your code must have one method: caesarSays(String, int) where the String represents the text to encrypt and the integer is the shift to apply. (There's no need for a decoding method; can you guess why?)

Good luck! And if you need any more information about a Cesar Cipher read the wiki article linked below or ask me for more help! (Preferably in person cause I can explain with pen and pad.  ) Happy coding!

Wikipedia Article: http://en.wikipedia.org/wiki/Caesar_Cipher
(Try to code it WITHOUT looking at the explanation on the wiki, first. If you're really, really stuck and I can't help guide you, you can take a peak.. but they kinda explain how to do it and cheating isn't fun!)